assessment/requirements.txt
Django==3.2.10
django-extensions==3.1.4
djangorestframework==3.11.2
jsonschema==4.5.1
psycopg2==2.9.3
Werkzeug==0.16.0
